requirements.txt
# Deep Learning Framework
torch>=1.13.0
torchvision>=0.14.0

# Computer Vision & Image Processing
opencv-python>=4.7.0
opencv-python-headless>=4.7.0
Pillow>=9.4.0
scikit-image>=0.20.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.2.0
hdbscan>=0.8.29

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Image Quality Assessment
lpips>=0.1.4

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
